+++ objectified-ui/public/WHATS_NEW.md
@@ -16,6 +16,7 @@ We continue to improve the platform based on your feedback with improvements and
   - Fixed additionalProperties in property editor to allow for types and schemas
 - Moved import to its own feature, completely rewrote the logic
 - Added ability to import from a file into classes (add a class from a file)
+- Removed Mermaid diagram display, changed to export option
 
 View our YouTube channel [here](https://www.youtube.com/@objectifieddev) for detailed tutorials and walkthroughs!